compressor/requirements.txt
Pillow>=8.4.0  # Use a compatible version
pika==1.2.0
